+++ classes/WpMatomo/Admin/AdminSettings.php
@@ -22,6 +22,7 @@ class AdminSettings {
 	const TAB_EXCLUSIONS  = 'exlusions';
 	const TAB_PRIVACY     = 'privacy';
 	const TAB_GEOLOCATION = 'geolocation';
+	const TAB_ADVANCED    = 'advanced';
 
 	/**
 	 * @var Settings
@@ -43,12 +44,14 @@ class AdminSettings {
 		$exclusions      = new ExclusionSettings( $this->settings );
 		$geolocation     = new GeolocationSettings( $this->settings );
 		$privacy         = new PrivacySettings();
+		$advanced        = new AdvancedSettings();
 		$setting_tabs    = array(
 			self::TAB_TRACKING   => $tracking,
 			self::TAB_ACCESS     => $access_settings,
 			self::TAB_PRIVACY    => $privacy,
 			self::TAB_EXCLUSIONS => $exclusions,
 			self::TAB_GEOLOCATION => $geolocation,
+			self::TAB_ADVANCED    => $advanced,
 		);
 
 		$setting_tabs = apply_filters( 'matomo_setting_tabs', $setting_tabs, $this->settings );